requirements.txt
streamlit
streamlit-autorefresh
pandas
gspread
google-auth
